feat(persona-engine): Phase F calibration — float scoring, softer weights, environment-aware checklist and questionnaire prompts
F.1 quality_score float + smoother outcome table
  scorers.py — QualityBreakdown.quality_score is now float (was int
  1..5). Outcome weights raised for the "degraded but observed"
  categories: partial 0.6→0.65, max_turns 0.4→0.5, abandoned/patience
  0.2→0.35, error 0.0→0.15. Weight rebalance when checklist present:
  outcome 0.6 / checklist 0.4  →  outcome 0.35 / checklist 0.65.
  Result clamped to [1.05, 4.95] to avoid absolute 1/5 saturation.
  main.py + adapters/job_store.py — JobResultResponse + JobState
  quality_score typed as float | None so pydantic accepts the new
  fractional values (pre-F this rejected with int_from_float).

F.2 questionnaire prompt — separate service quality from session outcome
  adapters/questionnaire_generator.py — new system-prompt guidance:
  tell the LLM to distinguish (A) real service defects, (B) environment
  constraints (wallet, signing), (C) agent capability limits, and only
  let (A) drive the score down. Default to 2~4 range for normal
  sessions; reserve 1 and 5 for strong evidence.

F.3 checklist adapter — UI accessibility fallback for env-gated items
  adapters/checklist_adapter.py — new system-prompt rule: for items
  that require real wallet signing / on-chain action, mark passed when
  the UI is observable and navigable even if the transaction itself
  can't be executed. memo must note "UI-only" in that case. blocked
  is now reserved for genuine session failure.

tests/test_scorers.py — rewritten for float scoring; new test for
  micro-differentiation between similar sessions (pre-F they often
  collapsed to the same integer). 12 tests total.

Measured effect on the N=20 persona × 21 human seed after re-running
the batch:
  jup.ag      rating_mean 2.2   → 2.8   (KS 0.70 → 0.57)
  magiceden   rating_mean 2.7   → 2.95  (KS 0.52 unchanged)
  Persona quality distribution opened up from integer {1, 4} to
  floats spanning 1.21, 2.56, 2.78, 3.22, 3.86, 4.51 — visible in
  the dashboard scatter as proper point spread.

Honest caveat worth writing down: Spearman ρ got *more* negative
(magiceden -0.38 → -0.69). Not because F regressed — the old noise
floor hid the real pattern. Personas in TEXT MODE still can't truly
observe dynamic dApps like jup.ag/magiceden, so their differentiated
scores reflect "persona profile priors" more than site quality. The
fix path is either (a) browser mode batches or (b) testing simpler
static targets. F has done the scoring-formula half of the job.

+++ apps/persona-engine/adapters/job_store.py
@@ -31,7 +31,7 @@ class JobState:
     session_id: str | None = None
     screenshot_paths: list[str] = field(default_factory=list)
     checklist_results: list[dict] = field(default_factory=list)
-    quality_score: int | None = None
+    quality_score: float | None = None
     quality_breakdown: dict[str, Any] = field(default_factory=dict)
     questionnaire_answers: list[dict] = field(default_factory=list)
     structured_report: dict[str, Any] = field(default_factory=dict)
